Seperates styling from components
