+++ app/Domits/src/screens/home/styles/Home.js
@@ -1,0 +1,12 @@
+import {StyleSheet} from "react-native";
+
+const styles = StyleSheet.create({
+    activityIndicatorContainer: {
+        width: '100%',
+        height: '100%',
+        justifyContent: 'center',
+        alignItems: 'center',
+    }
+})
+
+export default styles;
